# wtw_backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Web Toons Wala API",
    "version": "1.0.0",
    "description": "API documentation for Web Toons Wala application"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error message"
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "ERROR_CODE"
              },
              "message": {
                "type": "string",
                "example": "Detailed error message"
              },
              "timestamp": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        }
      },
      "Success": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Success message"
          },
          "data": {
            "type": "object",
            "description": "Response data"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/comic/getPaginatedComics": {
      "get": {
        "tags": ["Comics"],
        "summary": "Get paginated comics",
        "description": "Retrieve a paginated list of comics with optional filtering",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "genreId",
            "in": "query",
            "description": "Filter by genre ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search term",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort order for the results",
            "schema": {
              "type": "string"
            },
            "enum": ["newest", "oldest"]
          }
        ],
        "responses": {
          "200": {
            "description": "Comics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/comic/getComicDetail": {
      "get": {
        "tags": ["Comics"],
        "summary": "Get comic details",
        "description": "Retrieve details of a specific comic",
        "parameters": [
          {
            "name": "comicId",
            "in": "query",
            "description": "ID of the comic",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comic details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Comic not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/comic/getAdminDashboardCounts": {
      "get": {
        "tags": ["Comics"],
        "summary": "Get admin dashboard counts",
        "description": "Retrieve counts for admin dashboard",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard counts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/comic/addNewComic": {
      "post": {
        "tags": ["Comics"],
        "summary": "Add a new comic",
        "description": "Create a new comic with content",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "comicName": {
                    "type": "string",
                    "example": "Super Hero"
                  },
                  "description": {
                    "type": "string",
                    "example": "A story about a super hero"
                  },
                  "genreId": {
                    "type": "integer",
                    "example": 1
                  },
                  "comicContent": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "imageURL": {
                          "type": "string",
                          "example": "https://example.com/image.jpg"
                        },
                        "audioURL": {
                          "type": "string",
                          "example": "https://example.com/audio.mp3"
                        },
                        "length": {
                          "type": "number",
                          "example": 120.5
                        },
                        "sceneNumber": {
                          "type": "integer",
                          "example": 1
                        }
                      },
                      "required": [
                        "imageURL",
                        "audioURL",
                        "length",
                        "sceneNumber"
                      ]
                    }
                  }
                },
                "required": [
                  "comicName",
                  "description",
                  "genreId",
                  "comicContent"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comic created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/comic/updateComic": {
      "post": {
        "tags": ["Comics"],
        "summary": "Update a comic",
        "description": "Update an existing comic with content",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "comicId": {
                    "type": "integer",
                    "example": 1
                  },
                  "comicName": {
                    "type": "string",
                    "example": "Super Hero"
                  },
                  "description": {
                    "type": "string",
                    "example": "A story about a super hero"
                  },
                  "genreId": {
                    "type": "integer",
                    "example": 1
                  },
                  "comicContent": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "imageURL": {
                          "type": "string",
                          "example": "https://example.com/image.jpg"
                        },
                        "audioURL": {
                          "type": "string",
                          "example": "https://example.com/audio.mp3"
                        },
                        "length": {
                          "type": "number",
                          "example": 120.5
                        },
                        "sceneNumber": {
                          "type": "integer",
                          "example": 1
                        }
                      },
                      "required": [
                        "imageURL",
                        "audioURL",
                        "length",
                        "sceneNumber"
                      ]
                    }
                  }
                },
                "required": [
                  "comicId",
                  "comicName",
                  "description",
                  "genreId",
                  "comicContent"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comic updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Comic not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/comic/deleteComic": {
      "get": {
        "tags": ["Comics"],
        "summary": "Delete a comic",
        "description": "Delete an existing comic",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "comicId",
            "in": "query",
            "description": "ID of the comic to delete",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comic deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Comic not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}